+++ pkg/block/fetcher_test.go
@@ -24,6 +24,8 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	promtest "github.com/prometheus/client_golang/prometheus/testutil"
+	prommodel "github.com/prometheus/common/model"
+	"github.com/prometheus/prometheus/model/relabel"
 	"github.com/prometheus/prometheus/tsdb"
 	"github.com/thanos-io/objstore"
 	"github.com/thanos-io/objstore/objtesting"
@@ -1286,6 +1288,91 @@ func Test_ParseRelabelConfig(t *testing.T) {
     `), SelectorSupportedRelabelActions)
 	testutil.NotOk(t, err)
 	testutil.Equals(t, "unsupported relabel action: labelmap", err.Error())
+}
+
+func Test_ParseRelabelConfigWithTenants(t *testing.T) {
+	t.Run("empty content returns no config", func(t *testing.T) {
+		global, perTenant, err := ParseRelabelConfigWithTenants(nil, nil)
+		testutil.Ok(t, err)
+		testutil.Assert(t, global == nil, "expected nil global config")
+		testutil.Assert(t, perTenant == nil, "expected nil per-tenant config")
+
+		global, perTenant, err = ParseRelabelConfigWithTenants([]byte(""), nil)
+		testutil.Ok(t, err)
+		testutil.Assert(t, global == nil, "expected nil global config")
+		testutil.Assert(t, perTenant == nil, "expected nil per-tenant config")
+	})
+
+	t.Run("global list format", func(t *testing.T) {
+		content := []byte(`
+- source_labels: [__name__]
+  action: drop
+  regex: "global_drop_.*"
+`)
+		global, perTenant, err := ParseRelabelConfigWithTenants(content, nil)
+		testutil.Ok(t, err)
+		testutil.Assert(t, perTenant == nil, "expected nil per-tenant config for global format")
+		testutil.Equals(t, 1, len(global))
+		testutil.Equals(t, relabel.Drop, global[0].Action)
+		testutil.Equals(t, prommodel.LabelNames{"__name__"}, global[0].SourceLabels)
+	})
+
+	t.Run("per-tenant map format", func(t *testing.T) {
+		content := []byte(`
+tenant-a:
+  - source_labels: [__name__]
+    action: drop
+    regex: "tenant_a_drop_.*"
+tenant-b:
+  - source_labels: [__name__]
+    action: keep
+    regex: "tenant_b_keep_.*"
+`)
+		global, perTenant, err := ParseRelabelConfigWithTenants(content, nil)
+		testutil.Ok(t, err)
+		testutil.Assert(t, global == nil, "expected nil global config for per-tenant format")
+		testutil.Equals(t, 2, len(perTenant))
+		testutil.Equals(t, 1, len(perTenant["tenant-a"]))
+		testutil.Equals(t, relabel.Drop, perTenant["tenant-a"][0].Action)
+		testutil.Equals(t, relabel.Keep, perTenant["tenant-b"][0].Action)
+	})
+
+	t.Run("invalid global config fails validation", func(t *testing.T) {
+		// hashmod with no modulus is invalid.
+		content := []byte(`
+- action: hashmod
+  source_labels: [__name__]
+  target_label: shard
+`)
+		_, _, err := ParseRelabelConfigWithTenants(content, nil)
+		testutil.NotOk(t, err)
+	})
+
+	t.Run("invalid per-tenant config fails validation", func(t *testing.T) {
+		content := []byte(`
+tenant-a:
+  - action: hashmod
+    source_labels: [__name__]
+    target_label: shard
+`)
+		_, _, err := ParseRelabelConfigWithTenants(content, nil)
+		testutil.NotOk(t, err)
+	})
+
+	t.Run("unknown action is rejected", func(t *testing.T) {
+		content := []byte(`
+- action: bogus
+  source_labels: [__name__]
+`)
+		_, _, err := ParseRelabelConfigWithTenants(content, nil)
+		testutil.NotOk(t, err)
+	})
+
+	t.Run("malformed yaml is rejected", func(t *testing.T) {
+		content := []byte(`::: not valid yaml :::`)
+		_, _, err := ParseRelabelConfigWithTenants(content, nil)
+		testutil.NotOk(t, err)
+	})
 }
 
 func TestDeletionMarkFilter_HoldsOntoMarks(t *testing.T) {
